Changed API file structure
