+++ packages/core/src/api/formatConversions/simplifyBlocksRehypePlugin.ts
